Reworking to extract user information from token, using valid token as test data
Removing solution to store user login data including username, isAdmin and avatarUrl as separate localStorage element. This is because user information is contained in token so user field is now instantiated based on token value. If token is valid, extract user info and set as field. Otherwise, both token and user are null.

Replacing mock localStorage.getItem return value of 'token' with a valid test token. This is due to errors in tests that were throwing when parseJwt() was trying to parse a JWT from the string 'token', which was happening because parseJwt() is now called by extractUserFromToken().