Fixing bug in checking for invalid tokens
Previously, when a user set their own scigateway:token key pair in localStorage and it wasn't a valid JWT, parseJWT would crash due to an uncaught TypeError exception. This also led to a hanging cypress test (in reality the test had failed as the bug was not recognised). Fixing this by wrapping the parseJwt call inside extractUserFromToken in a try-catch statement.