Adding additional cypress test, refactoring code, removing unnecessary comments
Adding a test in login.spec.ts to check behaviour when an invalid or unsigned JWT is present in localStorage for scigateway:token. Ensures the user is not logged in by setting their own value in this field.

Refactoring extractUserFromToken to log an error message to the console in the event of the above taking place. Also removing unnecessary type specifying when declaring two variables in that same function.

Removing unnecessary comments describing the contents of a test token variable being used in unit tests for the auth providers

Adding comments in new cypress test for clarification