Adding "homepageUrl" field to settings.json to redirect to a specified route #678
This new field currently acts the exact same as startUrl but overrides it if both are specified. This allows for a route to be specified in one of the plugins to act as the homepage (this is designed for where the homepage route should be given).

The homepage is registered as a route in datagateway-dataview-settings.json and the homepageUrl "/datagateway" should be specified in settings.json. At the moment, the homepage does not work universally: the SciGateway logo is still present on the mainAppBar and clicking this still redirects to the original SciGateway homepage. A later commit will address this so that any navigations to / will redirect to "homepageUrl" if there is one.

Another possible plan for the future is to not have the homepage route appear in the navigation draw and have it primarily accessible through the logo in the mainAppBar. This would involve replacing the scigateway logo with the datagateway logo and it's up for debate if this should happen yet.

Currently unsure if all of this should work alongside startUrl or replace it altogether.