Now universally redirecting to new homepage #678
When a homepageUrl is set, all attempts to access / will redirect to that URL. If the URL is not a registered route, a 404 will be displayed. This will mean no homepage is available so if a homepage is set, it must be correctly configured.

Removing work from previous commit which amended the redirection of clicking on the SciGateway logo in the mainAppBar. This is now restored to its original.