Reducing spacing, centering text and using one dg logo in nav draw #737
This addresses the changes requested by Franz to try and eliminate some of the horizontal spacing between entries in the nav draw. The DataGateway logo now only appears once at the top of the drawer.

Still work to do as the datagateway logo is not referenced from the best place yet. Next commit will address that.