Ensuring logo appears every time, amending tests #737
By adding ? when referencing the image to display in the nav bar, we avoid the race condition/bug that can occur on a page refresh when not all the plugins are loaded. When this bug occurs, the <img> can't be sourced properly sometimes, leading to a page crash. This fixes that.

Amending tests to account for snapshot changes and the fact that nav bar items are now simplified (don't include a logo and alt text each), meaning some tests can be taken out. Added a couple in as well.