Redoing unit tests for amended login functionality #735
This addresses scenarios where the maintenance state can't be fetched or set, which would previously log the user out. It also does the same thing if the github code is invalid. From now on, this means the user remains in their current login state.

Tests and functionality are also being put back to their original behaviour for what to do if tokens expire. This should correctly log the user out.