Fixing dropdown not displaying on direct route #735
A bug was occurring whereby the dropdown authenticator selector was not appearing when /login was directly navigated to. This was due to the mnemonic not being read properly into the loginComponent. This is being fixed by using a hook to read this value whenever it is detected to change in state.