Fixing failing e2e test #735
The mnemonic was being overwritten on a failed login attempt using user/pass to an empty string. This caused an unexpected call to /authenticators, which failed.

To accommodate the time for the error message to display on a failed login attempt (as it can sometimes vary) an artificial wait of 1 second has been added to the test.