Adding test to bump coverage, removing if statement #735
Adding a test for LoginSelector to ensure that it changes the mnemonic in local LoginPage state when the user changes this in the dropdown menu.

Also removing an if statement in the githubAuthProvider that would call the logOut function if the user tries to log in and gets a forbidden error. In accordance with other changes in this PR, logOut should not be called in any login attempt.