/login not accessible if logged in, reverting login logic #735
This makes the login page inaccessible if the user is already logged in. This requires the user to explicitly log out before trying to login to another session, which is ultimately more straightforward.

This means logic that was added to keep the user logged in after unsuccessfully trying to login to another account is no longer required so will be reverted back to what's on master.