Adding Contact Us form for accessibility enquiries #990
This is handled by a Microsoft Form embedded into the page. The link for this form is set through public/settings.json using "contactUsAccessibilityFormUrl".

If a URL is not set in the settings, the form is replaced by a mailto email address. This email address is whatever value is set in accessibility-page.contact-info in res/default.json. This should be an ISIS email for ISIS deployment and DLS email for DLS deployment.