Making contact us form url a mandatory setting, combining similar unit tests
"contactUsAccessibilityFormUrl" is now a mandatory setting, although it can be an empty string. This documents it as something that should be included but if we don't have a working form for whatever reason, we can leave this setting blank to show the email address to contact instead. This change requires setting this field as an empty string in the initialState of SciGateway.

Also combining two unit tests for the accessibility page which test for the page root div itself and the contact us component. They were both being initialised in the same way and both elements should be included on the page.