Preventing app bar from opening nav drawer on failed login attempt #999
This was happening whenever the user would try and login and fail. If autoLogin is available, the autoLogin will login, which then triggers a useEffect in mainAppBar which causes the nav drawer to open.

This same hook causes the nav drawer to open whenever the user refreshes their browser on any other page. So I've disabled this behaviour on the login page to get the correct functionality to occur.