Homepage now loads by default at web root, showing navigation bar when not logged in #1049
The user's login state now has no effect on access to the homepage or display of the navigation drawer. This means that DLS now allows access to the homepage at the web root and the navigation drawer shows all web app routes (access to these is determined elsewhere). This is in comparison to how things used to work on DLS, whereby all nav drawer routes including the homepage were inaccessible until the user logged in.