Fixing bug with post-login referral #1049
Since this PR is related to SciGateway login changes, I thought I'd lump in this additional fix. This addresses the problem whereby the user would try to navigate to a protected resource, be redirected to login, login successfully and then be mistakenly directed to the homepage instead of their original page. This was previously working by storing the referrer in the login redirect state but this state is cleared on a page refresh, meaning we defaulted back to the homepage. We now store the referrer in localStorage and clear it when we don't need it.