Fixing original referrer issue, removing reliance on localStorage #1049
The reason the referrer was being wiped from state was because the user was logging in, which meant the login page becomes inaccessible and is thus replaced with logout. This removes the referrer from state. I am now changing this so we grab the referrer earlier in the process before redirecting. This removes the need to use localStorage